refactor: 기관 검색 훅으로 분리
